requirements.txt
requests>=2.28.0
numpy>=1.22.0
pandas>=1.4.0
matplotlib>=3.5.0
seaborn>=0.11.0
python-dotenv>=0.20.0
pytz>=2022.1
tqdm>=4.64.0